+++ src/main/java/roomescape/reservation/Reservation.java
@@ -1,6 +1,7 @@
-package roomescape.model;
+package roomescape.reservation;
 
 import java.time.LocalDate;
+import roomescape.reservationtime.ReservationTime;
 
 public record Reservation(Long id, String name, LocalDate date, ReservationTime time) {
 }
