+++ src/main/java/roomescape/reservationtime/ReservationTimeRepository.java
@@ -1,33 +1,30 @@
-package roomescape.repository;
+package roomescape.reservationtime;
 
 import java.time.LocalTime;
 import java.util.List;
 import org.springframework.stereotype.Repository;
-import roomescape.dao.ReservationTimeDao;
-import roomescape.model.ReservationTime;
 
 @Repository
-public class ReservationTimeRepository {
-
+class ReservationTimeRepository {
     private final ReservationTimeDao reservationTimeDao;
 
-    public ReservationTimeRepository(ReservationTimeDao reservationTimeDao) {
+    ReservationTimeRepository(ReservationTimeDao reservationTimeDao) {
         this.reservationTimeDao = reservationTimeDao;
     }
 
-    public List<ReservationTime> findAll() {
+    List<ReservationTime> findAll() {
         return reservationTimeDao.findAll();
     }
 
-    public ReservationTime save(LocalTime startAt) {
+    ReservationTime save(LocalTime startAt) {
         return reservationTimeDao.save(startAt);
     }
 
-    public void delete(Long id) {
+    void delete(Long id) {
         reservationTimeDao.delete(id);
     }
 
-    public ReservationTime findById(Long id) {
+    ReservationTime findById(Long id) {
         return reservationTimeDao.findById(id);
     }
 }
