+++ src/main/java/roomescape/reservationtime/ReservationTimeService.java
@@ -1,21 +1,23 @@
-package roomescape.service;
+package roomescape.reservationtime;
 
 import java.time.LocalTime;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.model.ReservationTime;
-import roomescape.repository.ReservationTimeRepository;
 
 @Service
 public class ReservationTimeService {
     private final ReservationTimeRepository reservationTimeRepository;
 
-    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
+    ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
         this.reservationTimeRepository = reservationTimeRepository;
     }
 
     public List<ReservationTime> findAll() {
         return reservationTimeRepository.findAll();
+    }
+
+    public ReservationTime findById(Long id) {
+        return reservationTimeRepository.findById(id);
     }
 
     public ReservationTime save(LocalTime startAt) {
